requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.3
mysql-connector-python==9.2.0
packaging==24.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
